src/rest/requirements.txt
Django==3.2
djangorestframework
pymongo
django-cors-headers
